Rename file && protect class
